Replacing DLibDetector with FaceDetector
