+++ GazeInference_WinCpp/packages.config
@@ -3,7 +3,7 @@
   <package id="Dlib-Native" version="19.22.0.1" targetFramework="native" />
   <package id="FaceLibrary-Native" version="1.0.0" targetFramework="native" />
   <package id="Microsoft.AI.DirectML" version="1.5.1" targetFramework="native" />
-  <package id="Microsoft.ML.OnnxRuntime" version="1.8.0" targetFramework="native" />
-  <package id="Microsoft.ML.OnnxRuntime.DirectML" version="1.8.0" targetFramework="native" />
+  <package id="Microsoft.ML.OnnxRuntime" version="1.8.1" targetFramework="native" />
+  <package id="Microsoft.ML.OnnxRuntime.DirectML" version="1.8.1" targetFramework="native" />
   <package id="OpenCV-Native" version="4.5.2.1" targetFramework="native" />
 </packages>
